chrome/test/data/xr/e2e_test_files/html/webxr_test_camera_access.html

<!--
Tests that AR camera access returns non-empty GLTexture.
-->
<html>
  <head>
    <link rel="stylesheet" type="text/css" href="../resources/webxr_e2e.css">
  </head>
  <body>
    <canvas id="webgl-canvas"></canvas>
    <script src="../../../../../../third_party/blink/web_tests/resources/testharness.js"></script>
    <script src="../resources/webxr_e2e.js"></script>
    <script src="../resources/webxr_boilerplate.js"></script>
    <script>var shouldAutoCreateNonImmersiveSession = false;</script>
    <script>var immersiveArSessionInit = { requiredFeatures: ['camera-access'] };</script>
    <script>
      setup({single_test: true});

      let numPosesFound = 0;
      let cameraImageTexture = null;
      let ranAtLeastOnce = false;

      function stepStartStoringCameraTexture(numCalls) {
        const webglCanvas = document.getElementById('webgl-canvas');
        const gl = webglCanvas.getContext('webgl', {
                xrCompatible: true
            });
        const sessionInfo = sessionInfos[sessionTypes.AR];
        const referenceSpace = sessionInfo.currentRefSpace;

        onARFrameCallback = (session, frame) => {
          const glBinding = new XRWebGLBinding(session, gl);
          const pose = frame.getViewerPose(referenceSpace);
          if (pose) {
            // Wait several frames to allow the image buffer to be populated by a camera
            // texture before attempting to get the camera image texture.
            if (numPosesFound >= 10) {
              for (let view of pose.cameraViews) {
                ranAtLeastOnce = true;
                // Used to test that multiple consecutive calls will all return non-null textures.
                for (let remainingCalls = numCalls; remainingCalls > 0; remainingCalls--) {
                  cameraImageTexture = glBinding.getCameraImage(frame, view);
                  if (cameraImageTexture === null) {
                    onARFrameCallback = null;
                    assert_unreached("XRWebGLBinding.getCameraImage(...) returned null texture.");
                  }
                }
              }
            }
            numPosesFound++;
          }

          if (numPosesFound > 10) {
            onARFrameCallback = null;
            assert_equals(ranAtLeastOnce, true);
            done();
          }
        };
      }

      function stepStartStoreAndDeleteCameraTexture() {
        const webglCanvas = document.getElementById('webgl-canvas');
        const gl = webglCanvas.getContext('webgl', {
                xrCompatible: true
            });
        const sessionInfo = sessionInfos[sessionTypes.AR];
        const referenceSpace = sessionInfo.currentRefSpace;

        onARFrameCallback = (session, frame) => {
          const glBinding = new XRWebGLBinding(session, gl);
          const pose = frame.getViewerPose(referenceSpace);
          if (pose) {
              for (let view of pose.cameraViews) {
                ranAtLeastOnce = true;
                cameraImageTexture = glBinding.getCameraImage(frame, view);
                gl.deleteTexture(cameraImageTexture);
              }
              numPosesFound++;
          }

          if (numPosesFound > 10) {
            onARFrameCallback = null;
            assert_equals(ranAtLeastOnce, true);
            done();
          }
        };
      }
    </script>
  </body>
</html>
